fix(headers.py):  fix bug in handling of [shift+tab]
The `sublime-rst-completion` Package binds [shift+tab] to the
headers::SmartFoldingCommand().  However, it was failing to honor the
conditions when it should have been running the `unindent` command
instead of inserting a '\t' character.  These rules are visible in the
Sublime Text default Key Bindings, where the below rules (top to bottom)
are processed in the Key Bindings bottom to top:

- '\n' appears in selected text (i.e. multiple lines are selected),
- text before cursor is zero-or-more tabs or spaces (see note below),
- settings.shift_tab_unindent == True

If any of the above conditions is true, then `unindent` command must be run.
Otherwise, it appropriate to insert a '\t' character as a response to [shift+tab].
(This assumes the user has not overridden the default [shift+tab] Key Bindings.

Note:  It is verified that Sublime Text runs the `unindent` command, even
when the cursor is at the beginning of the line and a non-blank character
is immediately to the right of it.  No change occurs, but the command
log shows it was run.  The code in this PR aligns with that behavior.

This PR fixes the erroneous behavior of the "smart_folding" comand.